requirements.txt
streamlit
pandas
python-dotenv
google-api-python-client
google-generativeai
scikit-learn
sentence-transformers
nltk
emoji
numpy
